server/requirements.txt
flask
flask-cors
flask-bcrypt
flask-sqlalchemy
uuid
python-dotenv
requests
psycopg2
boto3